% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSASmap.R
\name{all.isolated}
\alias{all.isolated}
\title{Util function all.isolated}
\usage{
\method{all}{isolated}(n)
}
\arguments{
\item{n}{network object}
}
\value{
logical
}
\description{
Test if all vertexes are isolated
}

