% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSASmapJS.R
\name{calNetwork}
\alias{calNetwork}
\title{Find nodes from edges}
\usage{
calNetwork(dirPath)
}
\arguments{
\item{dirPath}{Path to the folder that holds all SAS scripts}
}
\description{
Find nodes from edges
}

