% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{countDataSteps}
\alias{countDataSteps}
\title{Count number of data steps in a SAS file}
\usage{
countDataSteps(sasCode)
}
\arguments{
\item{sasCode}{Character string containing SAS code}
}
\description{
Count number of data steps in a SAS file
}
\examples{
sasFile <- system.file('examples/SAScode/MainAnalysis.SAS', package='sasMap')
sasCode <- loadSAS(sasFile)
countDataSteps(sasCode)
}
