% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{countProcSteps}
\alias{countProcSteps}
\title{Count number of proc steps in a SAS file}
\usage{
countProcSteps(sasCode)
}
\arguments{
\item{sasCode}{Character string containing SAS code}
}
\description{
Count number of proc steps in a SAS file
}
