% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{countStatements}
\alias{countStatements}
\title{Count statements in a SAS file}
\usage{
countStatements(sasCode)
}
\arguments{
\item{sasCode}{Character string containing SAS code}
}
\description{
Count statements in a SAS file
}
\examples{
sasFile <- system.file('examples/SAScode/MainAnalysis.SAS', package='sasMap')
sasCode <- loadSAS(sasFile)
countStatements(sasCode)
}
