% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawProcs.R
\name{drawProcs}
\alias{drawProcs}
\title{Draw the frequency of proc calls}
\usage{
drawProcs(procs)
}
\arguments{
\item{procs}{data.frame output from the listProcs function, i.e. counts by procs.}
}
\value{
\code{NULL}.
}
\description{
Draw a barplot based on the frequency.
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
out <- listProcs(parseSASfolder(sasDir)$Procs)
drawProcs(out)
}
}
\author{
Mango Solutions
}

