% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractLibrefs}
\alias{extractLibrefs}
\title{Extract Librefs from vector of SAS code}
\usage{
extractLibrefs(sasCode)
}
\arguments{
\item{sasCode}{}
}
\value{

}
\description{
Extract Librefs from vector of SAS code
}
