% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractMacroDefs}
\alias{extractMacroDefs}
\title{Extract macro definitions from a string of SAS code}
\usage{
extractMacroDefs(sasCode)
}
\arguments{
\item{sasCode}{SAS code}
}
\value{
Vector of macro definitions
}
\description{
Extract macro definitions from a string of SAS code
}
