% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractProcs}
\alias{extractProcs}
\title{Extract  procs from a vector of SAS code}
\usage{
extractProcs(sasCode)
}
\arguments{
\item{sasCode}{SAS code}
}
\description{
Extract  procs from a vector of SAS code
}
