% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSASmapJS.R
\name{findEdges}
\alias{findEdges}
\title{Extract edges dataset}
\usage{
findEdges(dirPath)
}
\arguments{
\item{dirPath}{Path to the folder that holds all SAS scripts}
}
\description{
Extract edges dataset
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
findEdges(sasDir)
}
}

