% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSASmapJS.R
\name{findNodes}
\alias{findNodes}
\title{Find nodes from edges}
\usage{
findNodes(edges, dirPath, sas.pattern = "\\\\.[Ss][Aa][Ss]$")
}
\arguments{
\item{edges}{The overall edges dataset including both natural and assessed edges}

\item{dirPath}{Path to the folder that holds all SAS scripts}

\item{sas.pattern}{SAS script extension}
}
\description{
Find nodes from edges
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
edges <- findEdges(sasDir)
nodes <- findNodes(edges, sasDir)
} 
}

