% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{loadSAS}
\alias{loadSAS}
\title{Load SAS Code in}
\usage{
loadSAS(sasPath)
}
\arguments{
\item{sasPath}{SAS file location}
}
\value{
Character string containing contents of the file
}
\description{
Load SAS Code in
}
\examples{
sasPath <- system.file('examples/SAScode/Macros/fun2.SAS', package='sasMap')
loadSAS(sasPath)
}
