% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSASfolder.R
\name{parseSASfolder}
\alias{parseSASfolder}
\title{Extract summary information of data steps and proc calls by parsing a SAS folder.}
\usage{
parseSASfolder(sasDir, output = "data.frame")
}
\arguments{
\item{sasDir}{Path to the folder that holds all SAS scripts}

\item{output}{Output should be data.frame or list? Default is data.frame.}
}
\value{
\code{data.frame} or \code{list}.
}
\description{
Extract summary information of data steps and proc calls by parsing a SAS folder.
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
sasCode <- parseSASfolder(sasDir)
write.csv(sasCode, 'sasCode HSBC.csv', row.names = FALSE)
}
}
\author{
Mango Solutions
}

