% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSASscript.R
\name{parseSASscript}
\alias{parseSASscript}
\title{Parse the SAS script}
\usage{
parseSASscript(txt, allFiles = sub("\\\\.sas$", "",
  casefold(basename(list.files(pattern = "\\\\.[Ss][Aa][Ss]$", recursive =
  TRUE)))), output = "data.frame")
}
\arguments{
\item{txt}{Path to a SAS script}

\item{allFiles}{Character vector indicating all sas scripts}

\item{output}{Output should be data.frame or list? Default is data.frame.}
}
\value{
\code{data.frame} or \code{character}.
}
\description{
Parse the SAS script
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
parseSASscript(txt=file.path(sasDir, 'MainAnalysis.SAS'))
}
}
\author{
Mango Solutions
}

