% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSASmap.R
\name{plotSASmap}
\alias{plotSASmap}
\title{Plot static network of SAS acripts}
\usage{
plotSASmap(networkobj, displayisolates = FALSE, displaylabels = TRUE,
  label.cex = 1, arrowhead.cex = 0.8, edge.lwd = 0.1,
  vertex.col = rep(2, networkobj$gal$n), pdffile, width = 5, height = 5,
  ...)
}
\arguments{
\item{networkobj}{The network object which can be produced by \code{\link{createNetwork}}.}

\item{displayisolates}{Should isolates be displayed?}

\item{displaylabels}{Should vertex labels be displayed?}

\item{label.cex}{Character expansion factor for label text.}

\item{arrowhead.cex}{Expansion factor for edge arrowheads.}

\item{edge.lwd}{line width scale for edges.}

\item{vertex.col}{color for vertices; may be given as a vector or a vertex attribute name, if vertices are to be of different colors.}

\item{pdffile}{The file name if output PDF file.}

\item{width}{Width of PDF.}

\item{height}{Height of PDF.}

\item{...}{other arguments e.g. \code{main} which is title, passed to \code{plot.network.default}}
}
\value{
\code{NULL}. A png or pdf file.
}
\description{
Plot is rendered in a graph window by Default. Alternatively, a pdf file can be created by specifying the file name.
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
net <- renderNetwork(sasDir)
plotSASmap(net)
plotSASmap(net, pdffile='static_sas_map.pdf', width=10, height=10)
}
}
\author{
Mango Solutions
}

