% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSASmapJS.R
\name{plotSASmapJS}
\alias{plotSASmapJS}
\title{Plot interactive network of SAS scripts}
\usage{
plotSASmapJS(dirPath)
}
\arguments{
\item{dirPath}{Path to the folder that holds all SAS scripts (> 1)}
}
\value{
\code{NULL}.
}
\description{
It creates basic force directed network graphics with depends on the networkD3 package.
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
plotSASmapJS(sasDir)
}
}
\author{
Mango Solutions
}

