% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderNetwork.R
\name{recurseIdx}
\alias{recurseIdx}
\title{Util function recurseIdx}
\usage{
recurseIdx(vidx, netm)
}
\arguments{
\item{vidx}{Index returned by which}

\item{netm}{Net matrix object}
}
\description{
Util function recurseIdx
}

