% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{removeAllComments}
\alias{removeAllComments}
\title{Remove single line comments from a character string containing SAS code}
\usage{
removeAllComments(sasCode)
}
\arguments{
\item{sasCode}{Character string containing SAS code}
}
\value{
Character string of SAS code with comments removed
}
\description{
This version of this function removes all comments.
}
