% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{removeSingleLineComments}
\alias{removeSingleLineComments}
\title{Remove single line comments from a vector of SAS statements}
\usage{
removeSingleLineComments(sasVec)
}
\arguments{
\item{sasVec}{Vector of SAS code}
}
\description{
Finds and removes all lines starting with "*"
}
