% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderNetwork.R
\name{renderNetwork}
\alias{renderNetwork}
\title{A Wrapper for the \code{\link{parseSASfolder}} and \code{\link{createNetwork}}.}
\usage{
renderNetwork(dirPath, sas.pattern = "\\\\.[Ss][Aa][Ss]$")
}
\arguments{
\item{dirPath}{basedir to analysis}

\item{sas.pattern}{the extention file name pattern}
}
\value{
network object
}
\description{
Also set vertex attribute for the toplevel scripts.
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
net <- renderNetwork(sasDir)
net
}
}
\author{
Mango Solutions
}

