% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawProcs.R
\name{drawProcs}
\alias{drawProcs}
\title{Draw the frequency of proc calls}
\usage{
drawProcs(procsCol)
}
\arguments{
\item{procsCol}{Character or factor indicating the Procs from the descSASscript output}
}
\value{
\code{NULL}.
}
\description{
It creates a static barplot using ggplot2.
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
drawProcs(parseSASfolder(sasDir)$Procs)
}
}
\author{
Mango Solutions
}

