% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawProcs.R
\name{drawProcs}
\alias{drawProcs}
\title{Draw the frequency of proc calls}
\usage{
drawProcs(dirPath)
}
\arguments{
\item{dirPath}{Path to single sas file or folder}
}
\value{
\code{NULL}
}
\description{
Draw a barplot based on the frequency.
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
drawProcs(sasDir)
}
}
\author{
Mango Solutions
}

