% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractLibrefs}
\alias{extractLibrefs}
\title{ExtractLibrefs}
\usage{
extractLibrefs(sasCode)
}
\arguments{
\item{sasCode}{PARAM_DESCRIPTION}
}
\value{
vector of char, where each element is a libref
}
\description{
self explanitory
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_c}},\code{\link[stringr]{str_extract}}
}
