% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawProcs.R
\name{listProcs}
\alias{listProcs}
\title{List proc calls}
\usage{
listProcs(filepath)
}
\arguments{
\item{filepath}{Path to single sas file or folder}
}
\value{
\code{data.frame}
}
\description{
This function reports counts of proc calls by group.
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
listProcs(sasDir)
}
}
\author{
Mango Solutions
}

