% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSASscript.R
\name{parseSASscript}
\alias{parseSASscript}
\title{Parse the SAS script}
\usage{
parseSASscript(txt, funs = list(Script = getScript_name, nLines = countLine,
  Procs = countProc, Data_step = countData_step, Macro_call = extractMacro_call,
  Macro_define = extractMacro_define), ...)
}
\arguments{
\item{txt}{Path to a SAS script}

\item{allFiles}{Character vector indicating all sas scripts}

\item{output}{Output should be data.frame or list? Default is data.frame.}

\item{txt}{Path to a SAS script}
}
\value{
\code{data.frame} or \code{character}.

\code{list}.
}
\description{
Parse the SAS script
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
parseSASscript(txt=file.path(sasDir, 'MainAnalysis.SAS'))
}
}
\author{
Mango Solutions

Mango Solutions
}

