% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSASscript.R
\name{parseSASscript}
\alias{parseSASscript}
\title{Parse the SAS script}
\usage{
parseSASscript(txt, funs = list(Script = getScript_name, nLines = countLine,
  Procs = countProc, Data_step = countData_step, Macro_call = extractMacro_call,
  Macro_define = extractMacro_define), output = "list", ...)
}
\arguments{
\item{txt}{Path to a SAS script}

\item{funs}{functions to calculate stats}

\item{output}{data.frame or list}
}
\value{
\code{list}.
}
\description{
Parse the SAS script
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
parseSASscript(txt=file.path(sasDir, 'MainAnalysis.SAS'))
}
}
\author{
Mango Solutions
}

