% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSASmapJS.R
\name{plotSASmapJS}
\alias{plotSASmapJS}
\title{Plot interactive network of SAS scripts}
\usage{
plotSASmapJS(funData)
}
\arguments{
\item{funData}{Parsed SAS code from multiple scripts}
}
\value{
\code{NULL}.
}
\description{
It creates basic force directed network graphics with depends on the networkD3 package.
}
\examples{
\dontrun{
library(sasMap)
sasDir <- system.file('examples/SAScode', package='sasMap')
sasCode <- parseSASfolder(sasDir, output = "list")
plotSASmapJS(funData=sasCode)
}
}
\author{
Mango Solutions
}

