% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{removeMultilineComments}
\alias{removeMultilineComments}
\title{Remove multiline comments from SAS code}
\usage{
removeMultilineComments(sasCode)
}
\arguments{
\item{sasCode}{Character string containing SAS code}
}
\description{
Remove multiline comments from SAS code
}
\examples{
sasFile <- system.file('examples/SAScode/Macros/Util2.SAS', package='sasMap')
sasCode <- loadSAS(sasFile)
removeMultilineComments(sasCode)
}
