% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summariseSASScript}
\alias{summariseSASScript}
\title{Summarise a SAS script}
\usage{
summariseSASScript(sasPath)
}
\arguments{
\item{sasPath}{Path to SAS script}
}
\description{
Summarise a SAS script
}
\examples{
sasPath <- system.file('examples/SAScode/Macros/fun2.SAS', package='sasMap')
summariseSASScript(sasPath)
}
